renamed and re-arranged components
